The right screen now shows the seconds digit in my large font. The font don't look quite right (I'll need to adjust it's conversion), but it works.
It too me a LONG time to figure out that the Arduino GFX library's drawBitmap expects PROGMEM data, not RAM data. It would have been really nice if that was documented. I understand why it's done (it avoids needing a RAM buffer), but what a mess that was to find.